+++ changelog.md
@@ -7,6 +7,7 @@
 
 - Add `extra-platforms` CLI command.
 - Add new `ALPINE`, `KALI`, `MANJARO`, `OPENWRT` platform definitions.
+- Add issue template with detection results reporting.
 
 ## [9.0.0 (2026-02-11)](https://github.com/kdeldycke/extra-platforms/compare/v8.0.0...v9.0.0)